requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
numpy==2.1.1
django==5.1.1
djangorestframework==3.15.2
joblib==1.4.2
openpyxl==3.1.5